add phone number

+++ app/src/main/java/com/example/lessonslist/domain/student/EditStudentItemPaymentBalanceUseCase.kt
@@ -4,7 +4,7 @@ package com.example.lessonslist.domain.student
 
 class EditStudentItemPaymentBalanceUseCase(private val studentListRepository: StudentListRepository) {
 
-    suspend fun editStudentItemPaymentBalance(studentItemId: Int, paymentBalance: Float) {
+    suspend fun editStudentItemPaymentBalance(studentItemId: Int, paymentBalance: Int) {
         studentListRepository.editStudentItemPaymentBalance(studentItemId, paymentBalance)
     }
 }
